+++ test/statements/if_test.dart
@@ -7,13 +7,13 @@ void main() {
 
     test('simple', () {
       final template = env.fromString('{% if true %}...{% endif %}');
-      expect(template.render(), equals('...'));
+      expect(template.renderMap(), equals('...'));
     });
 
     test('elif', () {
       final template = env.fromString('''{% if false %}XXX{% elif true
             %}...{% else %}XXX{% endif %}''');
-      expect(template.render(), equals('...'));
+      expect(template.renderMap(), equals('...'));
     });
 
     test('elif deep', () {
@@ -21,37 +21,37 @@ void main() {
           List.generate(999, (i) => '{% elif a == ${i + 1} %}${i + 1}').join() +
           '{% else %}x{% endif %}';
       final template = env.fromString(source);
-      expect(template.renderWr(a: 0), equals('0'));
-      expect(template.renderWr(a: 10), equals('10'));
-      expect(template.renderWr(a: 999), equals('999'));
-      expect(template.renderWr(a: 1000), equals('x'));
+      expect(template.render(a: 0), equals('0'));
+      expect(template.render(a: 10), equals('10'));
+      expect(template.render(a: 999), equals('999'));
+      expect(template.render(a: 1000), equals('x'));
     });
 
     test('else', () {
       final template =
           env.fromString('{% if false %}XXX{% else %}...{% endif %}');
-      expect(template.render(), equals('...'));
+      expect(template.renderMap(), equals('...'));
     });
 
     test('empty', () {
       final template = env.fromString('[{% if true %}{% else %}{% endif %}]');
-      expect(template.render(), equals('[]'));
+      expect(template.renderMap(), equals('[]'));
     });
 
     test('complete', () {
       final template =
           env.fromString('{% if a %}A{% elif b %}B{% elif c == d %}'
               'C{% else %}D{% endif %}');
-      expect(template.renderWr(a: 0, b: false, c: 42, d: 42.0), equals('C'));
+      expect(template.render(a: 0, b: false, c: 42, d: 42.0), equals('C'));
     });
 
     test('no scope', () {
       var template =
           env.fromString('{% if a %}{% set foo = 1 %}{% endif %}{{ foo }}');
-      expect(template.renderWr(a: true), equals('1'));
+      expect(template.render(a: true), equals('1'));
       template =
           env.fromString('{% if true %}{% set foo = 1 %}{% endif %}{{ foo }}');
-      expect(template.render(), equals('1'));
+      expect(template.renderMap(), equals('1'));
     });
   });
 }