+++ test/statements/imports_test.dart
@@ -18,32 +18,32 @@ void main() {
   group('include', () {
     test('context include', () {
       var template = env.fromString('{% include "header" %}');
-      expect(template.renderWr(foo: 42), equals('[42|23]'));
+      expect(template.render(foo: 42), equals('[42|23]'));
 
       template = env.fromString('{% include "header" with context %}');
-      expect(template.renderWr(foo: 42), equals('[42|23]'));
+      expect(template.render(foo: 42), equals('[42|23]'));
 
       template = env.fromString('{% include "header" without context %}');
-      expect(template.renderWr(foo: 42), equals('[|23]'));
+      expect(template.render(foo: 42), equals('[|23]'));
     });
 
     test('choise includes', () {
       var template = env.fromString('{% include ["missing", "header"] %}');
-      expect(template.renderWr(foo: 42), equals('[42|23]'));
+      expect(template.render(foo: 42), equals('[42|23]'));
 
       template = env
           .fromString('{% include ["missing", "missing2"] ignore missing %}');
-      expect(template.renderWr(foo: 42), equals(''));
+      expect(template.render(foo: 42), equals(''));
 
       template = env.fromString('{% include ["missing", "missing2"] %}');
-      expect(() => template.render(), throwsA(isA<TemplatesNotFound>()));
+      expect(() => template.renderMap(), throwsA(isA<TemplatesNotFound>()));
 
       // template names in error
       // https://github.com/pallets/jinja/blob/master/tests/test_imports.py#L122
 
       void testIncludes(Template template, Map<String, dynamic> context) {
         context['foo'] = 42;
-        expect(template.render(context), equals('[42|23]'));
+        expect(template.renderMap(context), equals('[42|23]'));
       }
 
       template = env.fromString('{% include ["missing", "header"] %}');
@@ -62,12 +62,12 @@ void main() {
 
     test('include ignore missing', () {
       var template = env.fromString('{% include "missing" %}');
-      expect(() => template.render(), throwsA(isA<TemplateNotFound>()));
+      expect(() => template.renderMap(), throwsA(isA<TemplateNotFound>()));
 
       for (var extra in ['', 'with context', 'without context']) {
         template =
             env.fromString('{% include "missing" ignore missing $extra %}');
-        expect(template.render(), equals(''));
+        expect(template.renderMap(), equals(''));
       }
     });
 
@@ -80,7 +80,7 @@ void main() {
       );
 
       final template = env.getTemplate('main');
-      expect(template.render(), equals('123'));
+      expect(template.renderMap(), equals('123'));
     });
 
     // TODO(test): unoptimized scopes