+++ test/statements/inheritance_test.dart
@@ -58,7 +58,7 @@ void main() {
     test('layout', () {
       final template = env.getTemplate('layout');
       expect(
-          template.render(),
+          template.renderMap(),
           equals('|block 1 from layout|block 2 from '
               'layout|nested block 4 from layout|'));
     });
@@ -66,7 +66,7 @@ void main() {
     test('level1', () {
       final template = env.getTemplate('level1');
       expect(
-          template.render(),
+          template.renderMap(),
           equals('|block 1 from level1|block 2 from '
               'layout|nested block 4 from layout|'));
     });
@@ -74,7 +74,7 @@ void main() {
     test('level2', () {
       final template = env.getTemplate('level2');
       expect(
-          template.render(),
+          template.renderMap(),
           equals('|block 1 from level1|nested block 5 from '
               'level2|nested block 4 from layout|'));
     });
@@ -82,7 +82,7 @@ void main() {
     test('level3', () {
       final template = env.getTemplate('level3');
       expect(
-          template.render(),
+          template.renderMap(),
           equals('|block 1 from level1|block 5 from level3|'
               'block 4 from level3|'));
     });
@@ -90,7 +90,7 @@ void main() {
     test('level4', () {
       final template = env.getTemplate('level4');
       expect(
-          template.render(),
+          template.renderMap(),
           equals('|block 1 from level1|block 5 from '
               'level3|block 3 from level4|'));
     });
@@ -109,7 +109,7 @@ void main() {
       );
 
       final template = env.getTemplate('c');
-      expect(template.render(), equals('--INTRO--|BEFORE|[(INNER)]|AFTER'));
+      expect(template.renderMap(), equals('--INTRO--|BEFORE|[(INNER)]|AFTER'));
     });
 
     test('working', () {
@@ -120,7 +120,7 @@ void main() {
     test('reuse blocks', () {
       final template = env.fromString('{{ self.foo() }}|{% block foo %}42'
           '{% endblock %}|{{ self.foo() }}');
-      expect(template.render(), equals('42|42|42'));
+      expect(template.renderMap(), equals('42|42|42'));
     });
 
     test('preserve blocks', () {
@@ -133,7 +133,7 @@ void main() {
       );
 
       final template = env.getTemplate('b');
-      expect(template.render(), equals('BA'));
+      expect(template.renderMap(), equals('BA'));
     });
 
     test('dynamic inheritance', () {
@@ -149,7 +149,7 @@ void main() {
 
       for (var i in [1, 2]) {
         expect(
-            template.renderWr(master: 'master$i'), equals('MASTER${i}CHILD'));
+            template.render(master: 'master$i'), equals('MASTER${i}CHILD'));
       }
     });
 
@@ -164,9 +164,9 @@ void main() {
       );
 
       final template = env.getTemplate('child');
-      expect(template.renderWr(master: 'master1'), equals('MASTER1CHILD'));
-      expect(template.renderWr(master: 'master2'), equals('MASTER2CHILD'));
-      expect(template.render(), equals('MASTER1CHILD'));
+      expect(template.render(master: 'master1'), equals('MASTER1CHILD'));
+      expect(template.render(master: 'master2'), equals('MASTER2CHILD'));
+      expect(template.renderMap(), equals('MASTER1CHILD'));
     });
 
     test('scoped block', () {
@@ -180,7 +180,7 @@ void main() {
       final template =
           env.fromString('{% extends "master.html" %}{% block item %}'
               '{{ item }}{% endblock %}');
-      expect(template.renderWr(seq: range(5)), equals('[0][1][2][3][4]'));
+      expect(template.render(seq: range(5)), equals('[0][1][2][3][4]'));
     });
 
     test('super in scoped block', () {
@@ -194,7 +194,7 @@ void main() {
       final template =
           env.fromString('{% extends "master.html" %}{% block item %}'
               '{{ super() }}|{{ item * 2 }}{% endblock %}');
-      expect(template.renderWr(seq: range(5)),
+      expect(template.render(seq: range(5)),
           equals('[0|0][1|2][2|4][3|6][4|8]'));
     });
 
@@ -202,8 +202,8 @@ void main() {
     // TODO: test fixed macro scoping bug
 
     test('double extends', () {
-      // TODO(fix): double extends
-      expect(() => Template(doubleextends), throwsException);
+      expect(() => Template(doubleextends),
+          throwsA(predicate((e) => e is TemplateError)));
     });
   });
 }