+++ test/tests_test.dart
@@ -7,29 +7,29 @@ void main() {
 
     test('No Paren For Arg 1', () {
       final template = env.fromString('{{ foo is sameas none }}');
-      expect(template.render({'foo': null}), equals('true'));
+      expect(template.renderMap({'foo': null}), equals('true'));
     });
 
     test('Defined', () {
       final template = env.fromString('{{ missing is defined }}|'
           '{{ true is defined }}');
-      expect(template.render(), equals('false|true'));
+      expect(template.renderMap(), equals('false|true'));
     });
 
     test('Even', () {
       final template = env.fromString('{{ 1 is even }}|{{ 2 is even }}');
-      expect(template.render(), equals('false|true'));
+      expect(template.renderMap(), equals('false|true'));
     });
 
     test('Odd', () {
       final template = env.fromString('{{ 1 is odd }}|{{ 2 is odd }}');
-      expect(template.render(), equals('true|false'));
+      expect(template.renderMap(), equals('true|false'));
     });
 
     test('Lower', () {
       final template =
           env.fromString('{{ "foo" is lower }}|{{ "FOO" is lower }}');
-      expect(template.render(), equals('true|false'));
+      expect(template.renderMap(), equals('true|false'));
     });
 
     test('Type Checks', () {
@@ -52,7 +52,7 @@ void main() {
           '{{ (10 ** 2) is number }}|'
           '{{ 3.14159 is number }}');
       expect(
-          template.render({'mydict': {}}),
+          template.renderMap({'mydict': {}}),
           equals('false|true|false|true|true|false|true|true|true|true|false|'
               'true|true|true|false|true|true|true'));
     });
@@ -61,13 +61,13 @@ void main() {
       final template = env.fromString('{{ [1, 2, 3] is sequence }}|'
           '{{ "foo" is sequence }}|'
           '{{ 42 is sequence }}');
-      expect(template.render(), equals('true|true|false'));
+      expect(template.renderMap(), equals('true|true|false'));
     });
 
     test('Upper', () {
       final template =
           env.fromString('{{ "FOO" is upper }}|{{ "foo" is upper }}');
-      expect(template.render(), equals('true|false'));
+      expect(template.renderMap(), equals('true|false'));
     });
 
     test('Equal To', () {
@@ -79,7 +79,7 @@ void main() {
           '{{ bar is eq ("ba" + "z") }}|'
           '{{ bar is eq bar }}|'
           '{{ bar is eq foo }}');
-      expect(template.render({'foo': 12, 'bar': 'baz'}),
+      expect(template.renderMap({'foo': 12, 'bar': 'baz'}),
           equals('true|false|true|true|false|true|true|false'));
     });
 
@@ -90,7 +90,7 @@ void main() {
           '{{ 2 is gt 1 }}|{{ 2 is gt 2 }}|{{ 2 is ge 2 }}|'
           '{{ 2 is ge 3 }}');
       expect(
-          template.render({'foo': 12, 'bar': 'baz'}),
+          template.renderMap({'foo': 12, 'bar': 'baz'}),
           equals('true|false|true|false|true|false|true|false|true|false|'
               'true|false'));
     });
@@ -98,19 +98,19 @@ void main() {
     test('Same As', () {
       final template = env.fromString('{{ foo is sameas false }}|'
           '{{ 0 is sameas false }}');
-      expect(template.render({'foo': false}), equals('true|false'));
+      expect(template.renderMap({'foo': false}), equals('true|false'));
     });
 
     test('Greater Than', () {
       final template = env.fromString('{{ 1 is greaterthan 0 }}|'
           '{{ 0 is greaterthan 1 }}');
-      expect(template.render(), equals('true|false'));
+      expect(template.renderMap(), equals('true|false'));
     });
 
     test('Less Than', () {
       final template = env.fromString('{{ 0 is lessthan 1 }}|'
           '{{ 1 is lessthan 0 }}');
-      expect(template.render(), equals('true|false'));
+      expect(template.renderMap(), equals('true|false'));
     });
 
     test('Multiple Test', () {
@@ -125,7 +125,7 @@ void main() {
           '"(us-east-1|ap-northeast-1)" '
           'or "stage" is matching "(dev|stage)" }}');
 
-      expect(template.render(), equals('false'));
+      expect(template.renderMap(), equals('false'));
       expect(
           items,
           equals([
@@ -144,7 +144,7 @@ void main() {
           '{{ 3 is in [1, 2] }}|'
           '{{ "foo" is in {"foo": 1}}}|'
           '{{ "baz" is in {"bar": 1}}}');
-      expect(template.render(),
+      expect(template.renderMap(),
           equals('true|true|false|true|false|true|false|true|false'));
     });
   });